+++ store/state.ts
@@ -1,5 +1,6 @@
 import { PaypalState } from '../types/PaypalState'
 
 export const state: PaypalState = {
-  trans: null
+  trans: null,
+  methods: null
 }